% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PatientExploreR-package.R
\docType{data}
\name{dataOntology}
\alias{dataOntology}
\title{Ontology concept associations}
\format{An object of class \code{data.table} (inherits from \code{data.frame}) with 3540241 rows and 6 columns.}
\usage{
dataOntology
}
\description{
3,540,241 obs. of 6 variables
}
\details{
\itemize{
  \item *concept_id: integer identifier
  \item *concept_name: string term
  \item *...
}
}
\keyword{datasets}
