% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PatientExploreR-package.R
\docType{package}
\name{PatientExploreR}
\alias{PatientExploreR}
\alias{PatientExploreR-package}
\alias{PatientExploreR-package}
\title{PatientExploreR}
\description{
Demonstrate PatientExploreR implementation
}
\author{
\strong{Maintainer}: Glicksberg Ben \email{neb255@gmail.com}

}
