% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runPatientExploreRApp.R
\name{runPatientExploreRApp}
\alias{runPatientExploreRApp}
\title{Run PatientExploreR flagship application example.}
\usage{
runPatientExploreRApp()
}
\description{
Launch demo application in `inst/examples/PatientExploreRApp`
}
\examples{
if (interactive()) {
   patientExploreR::runPatientExploreRApp()
}
}
